docker-flask/requirements.txt
Flask==0.10.1
redis-3.2.0
isort-4.3.9
